WIP on Eric-comment touchups for show/tell
