+++ server/zoteroIntegration/api.ts
@@ -6,7 +6,7 @@ app.get(
 	'/api/zoteroIntegration',
 	wrap((req, res) =>
 		getZoteroIntegration(req.user.id).then((integration) =>
-			res.status(201).json({ id: integration.id }),
+			integration ? res.status(201).json({ id: integration.id }) : res.status(404),
 		),
 	),
 );
