--- types/zoteroIntegration.ts
@@ -1,7 +1,0 @@
-export type ZoteroIntegration = {
-	id: string;
-	userId: string;
-	externalUsername: string;
-	externalUserId: string;
-	integrationDataOAuth1Id: string;
-};
